+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 v1.0.0
 ------
 
-* Initial release
+* Initial release with ZipTask and UnzipTask.